Initialize the project setting page
* Introduce a new command 'Open Project Settings' to open the new
  Project Setting Page.
* 'Configure Classpath' command will open the new page and focus
  the Classpath section.
* 'Configure Java Runtime' command will open the new page, focus
  the JDK tab under the Classpath section.
