--- src/classpath/assets/App.tsx
@@ -1,14 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license.
-
-import React from "react";
-import ClasspathConfigurationView from "./features/classpathConfiguration/ClasspathConfigurationView";
-
-export class App extends React.Component {
-
-    render() {
-        return (
-            <ClasspathConfigurationView />
-        );
-    }
-}
