--- src/classpath/assets/app/store.ts
@@ -1,11 +1,0 @@
-// Copyright (c) Microsoft Corporation. All rights reserved.
-// Licensed under the MIT license.
-
-import { configureStore } from "@reduxjs/toolkit";
-import classpathConfigurationViewReducer from "../features/classpathConfiguration/classpathConfigurationViewSlice";
-
-export default configureStore({
-  reducer: {
-    classpathConfig: classpathConfigurationViewReducer
-  },
-});
