+++ src/project-settings/assets/classpath/features/classpathConfigurationViewSlice.ts
@@ -1,0 +1,112 @@
+// Copyright (c) Microsoft Corporation. All rights reserved.
+// Licensed under the MIT license.
+
+import { createSlice, current } from "@reduxjs/toolkit";
+import _ from "lodash";
+import { ClasspathEntry } from "../../../handlers/classpath/types";
+
+export const classpathConfigurationViewSlice = createSlice({
+    name: "classpathConfig",
+    initialState: {
+      ui: {
+        activeTab: "",
+      },
+      data: {
+        activeVmInstallPath: [] as string[],
+        vmInstalls: [],
+        sources: [] as ClasspathEntry[][],
+        output: [] as string[],
+        libraries: [] as ClasspathEntry[][],
+      },
+      loadingState: false, // TODO: move to common?
+      exception: undefined,
+    },
+    reducers: {
+      updateActiveTab: (state, action) => {
+        state.ui.activeTab = action.payload;
+      },
+      initializeProjectsData: (state, action) => {
+        const projectNum = action.payload.projectsNum;
+        state.data.activeVmInstallPath = Array(projectNum).fill("");
+        state.data.sources = Array(projectNum).fill([]);
+        state.data.output = Array(projectNum).fill("");
+        state.data.libraries = Array(projectNum).fill([]);
+      },
+      listVmInstalls: (state, action) => {
+        state.data.vmInstalls = action.payload;
+      },
+      loadClasspath: (state, action) => {
+        const activeProjectIndex = action.payload.activeProjectIndex;
+        state.data.output[activeProjectIndex] = action.payload.output;
+        state.data.activeVmInstallPath[activeProjectIndex] = action.payload.activeVmInstallPath;
+        // Only update the array when they have different elements.
+        const currentSources = _.sortBy(current(state.data.sources), ["path", "output"]); // TODO: activeProjectIndex needed?
+        const newSources = _.sortBy(action.payload.sources, ["path", "output"]);
+        if (!_.isEqual(currentSources, newSources)) {
+          state.data.sources[activeProjectIndex] = action.payload.sources;
+        }
+
+        const currentLibs = _.sortBy(current(state.data.libraries), ["path"]);
+        const newLibs = _.sortBy(action.payload.libraries, ["path"]);
+        if (!_.isEqual(currentLibs, newLibs)) {
+          state.data.libraries[activeProjectIndex] = action.payload.libraries;
+        }
+      },
+      updateSource: (state, action) => {
+        const activeProjectIndex = action.payload.activeProjectIndex;
+        state.data.sources[activeProjectIndex] = _.uniqBy(action.payload.sources as ClasspathEntry[], "path");
+      },
+      setOutputPath: (state, action) => {
+        const activeProjectIndex = action.payload.activeProjectIndex;
+        state.data.output[activeProjectIndex] = action.payload.outputPath;
+      },
+      setJdks: (state, action) => {
+        const activeProjectIndex = action.payload.activeProjectIndex;
+        state.data.activeVmInstallPath[activeProjectIndex] = action.payload.activeVmInstallPath;
+        if (action.payload.vmInstalls &&
+              isDifferentStringArray(state.data.vmInstalls, action.payload.vmInstalls)) {
+          state.data.vmInstalls = action.payload.vmInstalls;
+        }
+      },
+      removeReferencedLibrary: (state, action) => {
+        const activeProjectIndex: number = action.payload.activeProjectIndex;
+        const removedIndex: number = action.payload.removedIndex;
+        if (removedIndex > -1 && removedIndex < state.data.libraries[activeProjectIndex].length) {
+          state.data.libraries[activeProjectIndex].splice(removedIndex, 1);
+        }
+      },
+      addLibraries: (state, action) => {
+        const activeProjectIndex = action.payload.activeProjectIndex;
+        let newLibs = state.data.libraries[activeProjectIndex];
+        newLibs.unshift(...action.payload.libraries);
+        newLibs = _.uniq(newLibs);
+        state.data.libraries[activeProjectIndex] = _.uniqBy(newLibs, "path");
+      },
+      catchException: (state, action) => {
+        state.exception = action.payload;
+      },
+      updateLoadingState: (state, action) => {
+        state.loadingState = action.payload;
+      },
+    },
+});
+
+function isDifferentStringArray(a1: string[], a2: string[]): boolean {
+  return !_.isEmpty(_.xor(a1, a2));
+}
+
+export const {
+  updateActiveTab,
+  initializeProjectsData,
+  listVmInstalls,
+  loadClasspath,
+  updateSource,
+  setOutputPath,
+  setJdks,
+  removeReferencedLibrary,
+  addLibraries,
+  catchException,
+  updateLoadingState,
+} = classpathConfigurationViewSlice.actions;
+
+export default classpathConfigurationViewSlice.reducer;
