+++ src/project-settings/assets/classpath/features/components/Libraries.tsx
@@ -5,29 +5,35 @@ import { Dispatch } from "@reduxjs/toolkit";
 import React, { useEffect, useState } from "react";
 import { useSelector, useDispatch } from "react-redux";
 import { removeReferencedLibrary, addLibraries } from "../classpathConfigurationViewSlice";
-import { onWillSelectLibraries } from "../../../utils";
+import { ClasspathRequest } from "../../../vscode/utils";
 import { VSCodeButton, VSCodeDataGrid, VSCodeDataGridCell, VSCodeDataGridRow, VSCodeDivider } from "@vscode/webview-ui-toolkit/react";
-import { ClasspathEntry, ClasspathEntryKind } from "../../../../types";
+import { ClasspathEntry, ClasspathEntryKind } from "../../../../handlers/classpath/types";
 
 const Libraries = (): JSX.Element => {
 
   const [hoveredRow, setHoveredRow] = useState<string | null>(null);
-
-  const libraries: ClasspathEntry[] = useSelector((state: any) => state.classpathConfig.libraries[state.classpathConfig.activeProjectIndex]);
+  const activeProjectIndex: number = useSelector((state: any) => state.commonConfig.ui.activeProjectIndex);
+  const libraries: ClasspathEntry[] = useSelector((state: any) => state.classpathConfig.data.libraries[activeProjectIndex]);
   const dispatch: Dispatch<any> = useDispatch();
 
   const handleRemove = (index: number) => {
-    dispatch(removeReferencedLibrary(index));
+    dispatch(removeReferencedLibrary({
+      activeProjectIndex,
+      removedIndex: index
+    }));
   };
 
   const handleAdd = () => {
-    onWillSelectLibraries();
+    ClasspathRequest.onWillSelectLibraries();
   };
 
   const onDidAddLibraries = (event: OnDidAddLibrariesEvent) => {
     const {data} = event;
-    if (data.command === "onDidAddLibraries") {
-      dispatch(addLibraries(data.jars));
+    if (data.command === "classpath.onDidAddLibraries") {
+      dispatch(addLibraries({
+        activeProjectIndex,
+        libraries:data.jars
+    }));
     }
   };
