+++ src/project-settings/assets/index.tsx
@@ -3,16 +3,11 @@
 
 import * as React from "react";
 import * as ReactDOM from "react-dom";
-import { Provider } from "react-redux";
-import { App } from "./App";
-import store from "./app/store";
-import "./style.scss";
+import App from "./mainpage/features/App";
 
 ReactDOM.render(
     <React.StrictMode>
-        <Provider store={store}>
-            <App />
-        </Provider>
+        <App />
     </React.StrictMode>,
     document.getElementById("content")
 );
