+++ src/project-settings/assets/utils.scss
@@ -38,3 +38,5 @@
     display: flex;
     align-items: center;
 }
+
+@import "~@vscode/codicons/dist/codicon.css";
